SGX: UDD: Use correct stride when TILER memory is used
This patch prevents a crash when TILER memory is being used. When
the framebuffer is using TILER the container physical stride of the
buffer should be used (depending on the bpp) instead of the 4K/8K
virtual stride since the SGX currently uses the physical address
of the buffer. This is a workaround to prevent such crash.

Change-Id: I14631743bb575b88cdc7fb2c0927d64c954eb744
Signed-off-by: Gustavo Diaz Prado <x0083741@ti.com>