OMAP4: PM: TEMP: Prevent l3init from idling
Since MMC driver is yet to be adapted to runtime PM and still uses direct
clock calls to enable/disbale module, its needed that the clockdomain is kept
under software control since clock framework does not take care of force
enabling the clockdomain when a clock is enabled. This is done only in hwmod
framework.

So the clockdomain data for l3init is changed to program to SW wakeup. This
will be eventually removed, when MMC is runtime adapted.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>