OMAP: PM: Fix bug while settting constraints.
The dummy-l3 device was created locally. So its address used
to differ in both request and release context.  The address was
different even if same device used to request the throughput-set
API.

Now the dummy-l3 device is created static, so that it gets
initialized only once and has a constant address.  With this
the clocks are released properly.

Signed-off-by: Avinash.H.M.<avinashhm@ti.com>