OMAP2+: PM: init_voltages: handle non compliant bootloaders
Bootloaders should in theory setup a frequency which is enabled in
OPP table. However, there can be mismatches, and we should try
both going lower in addition to the going higher to find
a lower match if bootloader boots up at a higher OPP than the
kernel thinks it should be allowed.

Reported-by: Colin Cross <ccross@google.com>
Signed-off-by: Nishanth Menon <nm@ti.com>