OMAP4: clock: Support divider selection for auxclks
On OMAP4 the auxclk nodes (part of SCRM) support both
divider as well as parent selection.
Supporting this requires splitting the existing nodes
(which support only parent selection) into
two nodes, one for parent and another for
divider selection.
The nodes for parent selection are named
auxclk*_src_ck and the ones for divider
selection as auxclk*_ck

Acked-by: Benoit Cousson<b-cousson@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>