OMAP4: clock: Define a virtual clock node to do l3 scaling
Scaling L3 as part of DVFS on OMAP4 involves multiple CORE DPLL
post dividers to be changed, and some in PER DPLL as well.
Hide all these from the DVFS/OPP layer by creating a
virtual clock node to handle all this internally as part
of the clock framework.

Also update the OMAP4 OPP table with the name of the virtual
clock.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>