OMAP4: PM: TPS: unlink VCORE3 and VMEM from APE for TWL6030
In 4460, the VMEM and VCORE3 resources are not populated.
So these resources should not be associated with any
power resource groups.

If associated, when the resource transition to OFF state,
the VMEM and VCORE3 get detected as shorted and in the
next wakeup, the TWL detects a exception (short cirucit
in previous cycle) and then powers down itself.

Since these are critical settings, we try to get subsequent
settings setup even if setting previous one failed.

Acked-by: Ambresh K <ambresh@ti.com>
[nm@ti.com: ported back to 3.0]
Signed-off-by: Nishanth Menon <nm@ti.com>
[rajeevk@ti.com: co-development]
Signed-off-by: Rajeev Kulkarni <rajeevk@ti.com>
Signed-off-by: Avinash Mahadeva <avinashhm@ti.com>