OMAP3+: ABB: core implementation
This patch implements the Adaptive Body-Bias LDO initialization routine
and the transition sequence needed during voltage scaling.

It also adds struct omap_abb_instance to struct voltagedomain and
populates the OMAP3 & OMAP4 voltage domain tables with their
corresponding ABB data.  Note that VDD_IVA on OMAP44xx does have an ABB
LDO, but this data is not populated in omap4_voltdm_iva due to
recommendations to bypass that LDO always.

Finally this patch adds the ABB transition sequence to the
vp_forceupdate method of voltage scaling.

There is a lot going on in this patch, but this is to keep from breaking
git bisect.

Signed-off-by: Mike Turquette <mturquette@ti.com>