OMAP3+: OPP: add ABB data to OPP tables
The operating mode of the Adaptive Body-Bias LDO maps directly to the
voltage of its voltage domain.  The two modes supported are bypass and
Forward Body-Bias (FBB).

This patch models this relationship by adding an abb_type paramter to
struct omap_volt_data and populates this type in the 3630, 4430 and 4460
OPP tables.

3430 does not have an ABB LDO at all, so it's abb_type values are
initialised to NONE for all domains.

{TBD: Information here is from pre-release Data Manual addendum}

Acked-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Mike Turquette <mturquette@ti.com>