OMAP4: PM: enable wakeup irq for PRCM and SYS_1N
Wake-up enable bits need to be set by each driver,
but some specific irq's do not have a driver as such,
and still have to be set as wakeup enable. This is the case
of PRCM and SYS_1N. For these irq's, we handle wakeup
enable bits on the suspend path.

Acked-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Signed-off-by: Deepak K<deepak.k@ti.com>