OMAP4: PM: Disable SR in idle
SmartReflex values are not defined for retention voltage. Further, with
SmartReflex enabled, if CPU enters lower c-states, FSM will try to bump
the voltage to current OPP's voltage for which it has entered c-state;
hence SmartReflex needs to be disabled for MPU, CORE and IVA voltage
domains in idle path before enabling auto retention voltage achievement
on the device.

Signed-off-by: Ambresh K <ambresh@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>