OMAP3+: VC: calculate the setup times dynamically
To scale to the voltage during states such as retention, the PMIC requires some
voltage stabilization time. This depends on the worst case ramp voltage and the
sysclk. Support is added to dynamically calculate the ramp counts and update
the appropriate registers due to the impact of Errata Id: i623 on OMAP4430
silicon alone.

Signed-off-by: Ambresh K <ambresh@ti.com>
[Ported to 3.0 and cleanup]
Signed-off-by: Avinash.H.M. <avinashhm@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>