OMAP4: PM: make omap4_pm_off_mode_enable flexible
Currently omap4_pm_off_mode_enable does not look at domain
capability and attempt the lowest power state for a requested
power state. make it more capable by introducing a search logic
which, given a lowest state we'd like to achieve, will search
through the domain's capability list and provide as output
the achievable state which we program hence.

This allows us to centralize the power domain configuration to a
central location.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>