OMAP4: Adding check in case CPU1 wakes up before CPU0 from device off or MPU OFF
It is imperative to enforce CPU0 to be ON before CPU1 can call any PPA HAL or
ROM HAL apis. Current handling of this scenario in kernel is not enough.
If CPU1 wakes up before CPU0, it should go to OFF again and await for CPU0 to
awaken him.

Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
santosh.shilimkar@ti.com: Fixed the code for GP devices.
Signed-off-by: Carlos Leija <cileija@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>