OMAP4: PM: skip going through the LP sleep sequence if conflict with DVFS
Currently we consider Core and MPU dvfs transitions independently.
In the case of core PD:
With frequency dependency on core domain, we have almost all devices
depending to scale core domain frequency scaling. Currently, we restrict
ourselves to checking just core or IVA domain before deciding the domains
should remain ON, instead we should be checking for all devices scaling.

In the case of MPU PD:
If we interrupt dvfs scaling executing on MPU for a different domain(e.g IVA),
we should let it complete rather than go through the motions of programming
and exiting the transition, and leaving the dvfs transition which had started
half done and mpu hitting idle state. In cases such as IVA dvfs, this might
tend to become even fatal as clk configurations need to go through syslink
on A9 when requested from Ducati - So even if A9 is not really busy, Ducati
could be waiting for the action to complete.

Currently we can create all kind of weird cases with DVFS Vs LP
use cases such as MPUSS successfully in OSWR but CORE in ON because
we interrupted SGX dvfs scale or something similar. Instead,
simplify sequencing by denying transition to LP if DVFS was caught
when we dont want to be interrupted.

We do this after clearing the previous states to allow the callers
to detect a failed transition.

Change-Id: Ib595f7393119e3b5bef89b5b3a827cf16060b274
Signed-off-by: Nishanth Menon <nm@ti.com>