OMAP4: PM: disable SR only for when VFSM comes into picture
Currently we are paranoid in disable of SR when core power
state is inactive or lower. We need to disable SR only when
VFSM comes into picture for going to retention voltage which
takes place only for retention or lower(OFF). So, change the
check for the same.

Side note: it makes sense to control the temp sensor accordingly
as well.

Change-Id: I54881922eb5c2f5c1efa42a1bd507bf2d6bc776a
Signed-off-by: Nishanth Menon <nm@ti.com>