OMAP4: PRM: improve cold reset logic
Cold reset logic currently could return back and invoke the
'BUG()' call which was supposed to be triggered if coldreset
was'nt functioning, even though the cold reset is eventually
triggered in h/w this gives a misconception that the cold
reset was'nt really functioning.

In OMAP4, the reset logic could take close to 100uSec before
the cold reset is percolated down from PRCM. Unfortunately,
we dont seem to have a status bit to check exactly when the
PRCM decides to send out cold reset signal to rest of h/w.

The delay is pretty much long enough for kernel messages to
be populated with the 'BUG()' call from pm44xx.c.

Further, it is mandatory that Power On reset bit is cleared
by s/w before cold reset is invoked. This is currently being
done by bootloader/PPA.

Change-Id: I886c420b6532e594b67da6a12122cac785ed6df6
Signed-off-by: Nishanth Menon <nm@ti.com>