OMAP3+: DVFS: ensure that current volt pointer reflects new volt pointer
At times when current_volt is same as new voltage, voltdm_scale is not
invoked. this can result in voltdm->curr_volt (which represents currently
active voltage data pointer not to reflect the exact voltage data structure.

So be sure to update the voltdm->curr_volt with exactly the same volt data
pointer as the new voltage is.

Reported-by: Honda Kenji <x0092217@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>